게시글 쓰기

+++ app/src/main/java/com/example/myapplication/Home/HomePostData.kt
@@ -6,7 +6,7 @@ data class HomePostData (
     val post_image: String = "",
     val post_place_name: String = "",
     val post_address: String = "",
-    val post_time: Int = 0, 
+    val post_time: Long = 0L,
     val post_rest_time: Int = 0,
     val post_price: String = "",
     val post_original_price: String = ""
